1.删除许多无法通过的算法 2.新增C语言课程作业

+++ LeetCode/美丽塔I/main.cpp
@@ -28,7 +28,7 @@ long long maximumSumOfHeights(vector<int> &maxHeights)
             maxHeights[i + 1] = maxHeights[i];
     }
 
-    int count = 0;
+     long long int count = 0;
     for (int i : maxHeights)
     {
         count += i;
